requirements.txt
flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0
openai==1.12.0
websockets==11.0.3
flask-sock==0.6.0
asyncio==3.4.3
requests==2.31.0 